refactor(core): extract routing and lb test modules
